third_party/WebKit/LayoutTests/http/tests/serviceworker/registration.html

<!DOCTYPE html>
<title>Service Worker: Registration</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script>
(function() {
    var t = async_test('Registering normal pattern');
    t.step(function() {
        navigator.serviceWorker.register(
            'resources/registration-worker.js',
            {scope: '/*'}
        ).then(
            t.step_func(function(worker) {
                assert_true(!!worker, 'Successfully registered.');
                t.done();
            }),
            t.step_func(function(reason) {
                assert_unreached('Registration should succeed, but failed: ' + reason.name);
            })
        );
    });
}());

(function() {
    var t = async_test('Registering scope outside domain');
    t.step(function() {
        navigator.serviceWorker.register(
            'resources/registration-worker.js',
            {scope: 'http://example.com/*'}
        ).then(
            t.step_func(function(worker) {
                assert_unreached('Registration scope outside domain should fail.');
            }),
            t.step_func(function(reason) {
                assert_equals(reason.name, 'SecurityError',
                              'Registration scope outside domain should fail with SecurityError.');
                t.done();
            })
        );
    });
}());

(function() {
    var t = async_test('Registering script outside domain');
    t.step(function() {
        navigator.serviceWorker.register(
            'http://example.com/worker.js'
        ).then(
            t.step_func(function(worker) {
                assert_unreached('Registration script outside domain should fail.');
            }),
            t.step_func(function(reason) {
                assert_equals(reason.name, 'SecurityError',
                             'Registration script outside domain should fail with SecurityError.');
                t.done();
            })
        );
    });
}());

</script>
